Refactoring of the App
I decided to add totaly new functionality to the app which is a modifying particular corner of our rectangle. To do this I changed completly the RectangleView.swift and I used GeometryReader to this. Apart from that I added a lot of new TextFields (one for every corner and one for the text on the rectangle). Right now I'm thinking about getting rid of the TurnButtonView.swift, because I don't see any need in this feature.